完成分配工作，实现默认到非默认到默认的状态切换； 阈值控制通过；

+++ code-api/src/main/java/com/hivescm/code/dto/CodeResult.java
@@ -24,7 +24,7 @@ public class CodeResult implements Serializable{
 	/**
 	 * 规则ID
 	 */
-	private Long ruleId;
+	private Integer ruleId;
 
 	public CodeResult() {
 	}
@@ -37,11 +37,11 @@ public class CodeResult implements Serializable{
 		this.code = code;
 	}
 
-	public Long getRuleId() {
+	public Integer getRuleId() {
 		return ruleId;
 	}
 
-	public void setRuleId(Long ruleId) {
+	public void setRuleId(Integer ruleId) {
 		this.ruleId = ruleId;
 	}
